refactor rules-of-analyse.js 🥝 ..
