+++ src/utils.js
@@ -34,4 +34,26 @@ const searchResult = (password) => ([
 ]);
 
 
-module.exports = { errorType, display, searchResult, espaceChart }
+// rules-of-analyse.js
+const rules = (TailleAnalysis ) => {
+  let rules = {};
+
+  if (TailleAnalysis > 9) {
+    rules = { 
+      firstRule : parseInt( TailleAnalysis * 0.3 ), 
+      secondRule : parseInt( TailleAnalysis * 0.1 ),
+      thirdRule : parseInt( TailleAnalysis * 0.1) 
+    };
+  }else{
+    rules = { 
+      firstRule : parseInt( TailleAnalysis * 0.3 ), 
+      secondRule : parseInt( TailleAnalysis * 0.2 ),
+      thirdRule : parseInt( TailleAnalysis * 0.2)
+    };
+  };
+
+  return rules ;
+};
+
+
+module.exports = { errorType, display, searchResult, espaceChart, rules }
